+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/main/java/com/swirlds/platform/test/event/source/AbstractEventSource.java
@@ -16,9 +16,9 @@
 
 package com.swirlds.platform.test.event.source;
 
-import static com.swirlds.platform.test.event.EventUtils.integerPowerDistribution;
-import static com.swirlds.platform.test.event.EventUtils.staticDynamicValue;
 import static com.swirlds.platform.test.event.RandomEventUtils.randomEventWithTimestamp;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.integerPowerDistribution;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.staticDynamicValue;
 
 import com.swirlds.common.system.NodeId;
 import com.swirlds.common.test.fixtures.TransactionGenerator;
