+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/consensus/ConsensusTestDefinitions.java
@@ -23,11 +23,11 @@ import static com.swirlds.platform.test.consensus.ConsensusUtils.applyEventsToCo
 import static com.swirlds.platform.test.consensus.ConsensusUtils.buildSimpleConsensus;
 import static com.swirlds.platform.test.consensus.ConsensusUtils.isRestartConsensusEquivalent;
 import static com.swirlds.platform.test.consensus.ConsensusUtils.testConsensus;
-import static com.swirlds.platform.test.event.EventUtils.createSignedState;
-import static com.swirlds.platform.test.event.EventUtils.getEventMap;
-import static com.swirlds.platform.test.event.EventUtils.getLastGenerationInState;
-import static com.swirlds.platform.test.event.EventUtils.integerPowerDistribution;
 import static com.swirlds.platform.test.event.source.EventSourceFactory.newStandardEventSources;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.createSignedState;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.getEventMap;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.getLastGenerationInState;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.integerPowerDistribution;
 import static com.swirlds.platform.test.graph.OtherParentMatrixFactory.createBalancedOtherParentMatrix;
 import static com.swirlds.platform.test.graph.OtherParentMatrixFactory.createCliqueOtherParentMatrix;
 import static com.swirlds.platform.test.graph.OtherParentMatrixFactory.createPartitionedOtherParentAffinityMatrix;
@@ -54,7 +54,6 @@ import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.internal.SignedStateLoadingException;
 import com.swirlds.platform.state.signed.SignedState;
 import com.swirlds.platform.test.SimpleEventGenerator;
-import com.swirlds.platform.test.event.EventUtils;
 import com.swirlds.platform.test.event.TestSequence;
 import com.swirlds.platform.test.event.emitter.PriorityEventEmitter;
 import com.swirlds.platform.test.event.emitter.StandardEventEmitter;
@@ -64,6 +63,7 @@ import com.swirlds.platform.test.event.source.EventSourceFactory;
 import com.swirlds.platform.test.event.source.ForkingEventSource;
 import com.swirlds.platform.test.event.source.StandardEventSource;
 import com.swirlds.platform.test.fixtures.event.DynamicValue;
+import com.swirlds.platform.test.fixtures.event.EventUtils;
 import com.swirlds.platform.test.fixtures.event.IndexedEvent;
 import java.io.IOException;
 import java.nio.file.Path;
