+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/consensus/EventEmitterTests.java
@@ -16,9 +16,9 @@
 
 package com.swirlds.platform.test.consensus;
 
-import static com.swirlds.platform.test.event.EventUtils.areEventListsEquivalent;
-import static com.swirlds.platform.test.event.EventUtils.areGenerationNumbersValid;
-import static com.swirlds.platform.test.event.EventUtils.isEventOrderValid;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.areEventListsEquivalent;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.areGenerationNumbersValid;
+import static com.swirlds.platform.test.fixtures.event.EventUtils.isEventOrderValid;
 import static com.swirlds.test.framework.TestQualifierTags.TIME_CONSUMING;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
